Started implementing users system.

+++ Backend/routes/routes_comments.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, jsonify, request, abort
-from flask_jwt_extended import jwt_required
+from flask_login import current_user, login_required
 from models.comments import Comment
 from __init__ import db, limiter
 from dotenv import load_dotenv
@@ -17,44 +17,16 @@ comments_bp = Blueprint("comments", __name__)
 load_dotenv(os.path.normpath("../.env"))
 
 
-def send_email(code: int, email: str):
-    try:
-        EMAIL = os.getenv("EMAIL")
-        PASSWORD = os.getenv("PASSWORD")
-        yag = yagmail.SMTP(EMAIL, PASSWORD)
-        contents = [
-            f"""
-            <html>
-                <body>
-                    <p>Dragă utilizator,</p>
-                    <p>Mai jos vei găsi codul unic necesar pentru a posta comentariul tău:</p>
-                    <p><b>Codul tău: {code}</b></p>
-                    <p>Te rugăm să introduci acest cod în câmpul corespunzător pe site pentru a continua.</p>
-                    <p>Dacă nu ai solicitat acest cod, te rog să ignori acest e-mail sau să ne contactezi pentru asistență.</p>
-                    <p>O zi excelentă,</p>
-                    <p>Echipa Feedback IESC</p>
-                </body>
-            </html>
-            """
-        ]
-        yag.send(email, "Codul tău pentru Feedback IESC", contents)
-        # logger.info(f"Email sent to {email} with code {code}.")
-    except Exception as e:
-        logger.error(
-            f"An error has occured while sending an email to {email}.\n {str(e)}"
-        )
-        return e
-
-
 def check_comments_number(email: str, subject_id: int) -> bool:
     try:
-        with db.session.begin():
-            existing_comments = (
-                db.session.query(Comment)
-                .filter_by(email=email, subject_id=subject_id)
-                .count()
-            )
-        return existing_comments < 2
+        existing_comments = (
+            db.session.query(Comment)
+            .filter_by(email=email, subject_id=subject_id)
+            .count()
+        )
+        if existing_comments < 5:
+            return True
+        return False
     except exc.SQLAlchemyError as e:
         logger.error(
             f"An error has occured while checking the number of comments.\n{str(e)}"
@@ -63,47 +35,44 @@ def check_comments_number(email: str, subject_id: int) -> bool:
 
 
 @comments_bp.route("/comments", methods=["POST"])
-@limiter.limit("50 per minute")
+@login_required
 def create_comment():
     try:
-        email = clean(request.form["email"])
-        code = randint(100000, 999999)
-        comment = ""
-        is_like = -1
-        is_anonymous = True
-        subject_id = clean(request.form["subject_id"])
-        grade = -1
+        email = current_user.email
+        comment = clean(request.form.get("comment"))
+        is_like = int(clean(request.form["is_like"]))
+        anonymous_choice = clean(request.form["is_anonymous"]).lower()
+        is_anonymous = True if anonymous_choice == "true" else False
+        subject_id = int(clean(request.form["subject_id"]))
+        grade = int(clean(request.form["grade"]))
         timestamp = datetime.now()
-        sentiment = -2
     except KeyError as e:
         logger.error(f"An error has occured: missing key in request parameters.\n {e}")
         abort(400, f"An error has occured: missing key in request parameters.")
-    if not check_comments_number(email, subject_id):
-        abort(
-            400,
-            f"User with email {email} has posted too many comments for this subject.",
-        )
-    send_email(code, email)
-    new_comment = Comment(
-        email,
-        code,
-        comment,
-        is_like,
-        is_anonymous,
-        subject_id,
-        grade,
-        timestamp,
-        sentiment,
-    )
+
     try:
-        with db.session.begin():
-            db.session.add(new_comment)
-            db.session.commit()
-            # logger.info("New comment intent added to database")
-            return {"response": "New comment intent added to database"}, 200
+        if not check_comments_number(email, subject_id):
+            abort(
+                400,
+                f"User with email {email} has posted too many comments for this subject.",
+            )
+
+        new_comment = Comment(
+            email,
+            comment,
+            is_like,
+            is_anonymous,
+            subject_id,
+            grade,
+            timestamp,
+        )
+        db.session.add(new_comment)
+        db.session.commit()
+        return {"response": "New comment added to database"}, 200
     except exc.SQLAlchemyError as e:
+        db.session.rollback()
         logger.error(f"An error has occured while adding object to the database.\n {e}")
-        return abort(500, f"An error has occured while adding object to the database.")
+        abort(500, f"An error has occured while adding object to the database.")
 
 
 @comments_bp.route("/comments", methods=["GET"])
@@ -121,7 +90,6 @@ def get_comments():
                         "is_like": comment.is_like,
                         "timestamp": comment.datetime.strftime("%d %b %Y"),
                         "email": "Anonim",
-                        "sentiment": comment.sentiment,
                     }
                     if not comment.is_anonymous:
                         comment_dict["email"] = comment.email
@@ -157,7 +125,6 @@ def get_comments_by_id(subject_id):
                     "is_like": comment.is_like,
                     "timestamp": comment.datetime.strftime("%d %b %Y"),
                     "email": "Anonim",
-                    "sentiment": comment.sentiment,
                 }
                 if not comment.is_anonymous:
                     comment_dict["email"] = comment.email
@@ -182,68 +149,24 @@ def get_comments_by_id(subject_id):
         abort(500, f"An error has occured while retrieving comments.")
 
 
-@comments_bp.route("/comments", methods=["PUT"])
-@limiter.limit("50 per minute")
-def update_comment():
-    try:
-        email = clean(request.form["email"])
-        new_comment = clean(request.form["new_comment"])
-        sent_code = clean(request.form["code"])
-        is_like = clean(request.form["is_like"])
-        subject_id = clean(request.form["subject_id"])
-        new_grade = int(clean(request.form["new_grade"]))
-        anonymous_choice = clean(request.form["is_anonymous"]).lower()
-        is_anonymous = True if anonymous_choice == "true" else False
-        if new_grade not in range(1, 11) and new_grade != -1:
-            raise KeyError
-    except KeyError as e:
-        logger.error(f"An error has occured: missing key in request parameters.\n {e}")
-        abort(400, f"An error has occured: missing key in request parameters.")
-    try:
-        with db.session.begin():
-            affected_rows = (
-                db.session.query(Comment)
-                .filter_by(
-                    email=email, code=sent_code, comment="", subject_id=subject_id
-                )
-                .update(
-                    {
-                        "comment": new_comment,
-                        "is_like": is_like,
-                        "is_anonymous": is_anonymous,
-                        "grade": new_grade,
-                        "datetime": datetime.now(),
-                    }
-                )
-            )
-            if affected_rows > 0:
-                db.session.commit()
-                logger.info(f"Comment insertion finalized.")
-                return {"response": f"Comment insertion finalized."}, 200
-            else:
-                logger.warning(f"No comment to finalize insertion.")
-                return abort(404, f"No comment to finalize insertion.")
-    except exc.SQLAlchemyError as e:
-        logger.error(f"An error has occured while updating object.\n {e}")
-        return abort(500, f"An error has occured while updating object.")
-
-
 @comments_bp.route("/comments/<int:comment_id>", methods=["DELETE"])
-@jwt_required()
+@login_required
 @limiter.limit("50 per minute")
 def delete_comment(comment_id):
-    try:
-        affected_rows = db.session.query(Comment).filter_by(id=comment_id).delete()
-        if affected_rows > 0:
-            db.session.commit()
-            logger.info(f"Comment with ID={comment_id} deleted")
-            return {"response": f"Comment with ID={comment_id} deleted"}, 200
-        else:
-            logger.warning(f"No comment with ID={comment_id} to delete")
-            return abort(404, f"No comment with ID={comment_id} to delete")
-    except exc.SQLAlchemyError as e:
-        logger.error(f"An error has occured while updating object.\n {e}")
-        return abort(500, f"An error has occured while updating object.")
+    if current_user.user_type == 0:
+        try:
+            affected_rows = db.session.query(Comment).filter_by(id=comment_id).delete()
+            if affected_rows > 0:
+                db.session.commit()
+                logger.info(f"Comment with ID={comment_id} deleted")
+                return {"response": f"Comment with ID={comment_id} deleted"}, 200
+            else:
+                logger.warning(f"No comment with ID={comment_id} to delete")
+                return abort(404, f"No comment with ID={comment_id} to delete")
+        except exc.SQLAlchemyError as e:
+            logger.error(f"An error has occured while updating object.\n {e}")
+            return abort(500, f"An error has occured while updating object.")
+    abort(401, "Not authorized.")
 
 
 @comments_bp.route("/nr_likes/<int:subject_id>", methods=["GET"])
